+++ compiler/frontend/ast_exp_handle_external.ml
@@ -43,7 +43,7 @@ let handle_external loc (x : string) : Parsetree.expression =
         Ast_external_mk.local_external_apply loc ~pval_prim:["#raw_expr"]
           ~pval_type:
             (Ast_helper.Typ.arrow ~arity:(Some 1)
-               {lbl = Nolabel; typ = Ast_helper.Typ.any ()}
+               {attrs = []; lbl = Nolabel; typ = Ast_helper.Typ.any ()}
                (Ast_helper.Typ.any ()))
           [str_exp];
     }
@@ -72,7 +72,7 @@ let handle_debugger loc (payload : Ast_payload.t) =
     Ast_external_mk.local_external_apply loc ~pval_prim:["%debugger"]
       ~pval_type:
         (Ast_helper.Typ.arrow ~arity:(Some 1)
-           {lbl = Nolabel; typ = Ast_helper.Typ.any ()}
+           {attrs = []; lbl = Nolabel; typ = Ast_helper.Typ.any ()}
            (Ast_literal.type_unit ()))
       [Ast_literal.val_unit ~loc ()]
   | _ ->
@@ -99,7 +99,7 @@ let handle_raw ~kind loc payload =
         Ast_external_mk.local_external_apply loc ~pval_prim:["#raw_expr"]
           ~pval_type:
             (Ast_helper.Typ.arrow ~arity:(Some 1)
-               {lbl = Nolabel; typ = Ast_helper.Typ.any ()}
+               {attrs = []; lbl = Nolabel; typ = Ast_helper.Typ.any ()}
                (Ast_helper.Typ.any ()))
           [exp];
       pexp_attributes =
@@ -128,11 +128,16 @@ let handle_ffi ~loc ~payload =
       let unit = Ast_literal.type_unit ~loc () in
       let rec arrow ~arity =
         if arity = 0 then
-          Ast_helper.Typ.arrow ~arity:None ~loc {lbl = Nolabel; typ = unit} any
+          Ast_helper.Typ.arrow ~arity:None ~loc
+            {attrs = []; lbl = Nolabel; typ = unit}
+            any
         else if arity = 1 then
-          Ast_helper.Typ.arrow ~arity:None ~loc {lbl = Nolabel; typ = any} any
+          Ast_helper.Typ.arrow ~arity:None ~loc
+            {attrs = []; lbl = Nolabel; typ = any}
+            any
         else
-          Ast_helper.Typ.arrow ~loc ~arity:None {lbl = Nolabel; typ = any}
+          Ast_helper.Typ.arrow ~loc ~arity:None
+            {attrs = []; lbl = Nolabel; typ = any}
             (arrow ~arity:(arity - 1))
       in
       match !is_function with
@@ -152,7 +157,7 @@ let handle_ffi ~loc ~payload =
           Ast_external_mk.local_external_apply loc ~pval_prim:["#raw_expr"]
             ~pval_type:
               (Ast_helper.Typ.arrow ~arity:(Some 1)
-                 {lbl = Nolabel; typ = Ast_helper.Typ.any ()}
+                 {attrs = []; lbl = Nolabel; typ = Ast_helper.Typ.any ()}
                  (Ast_helper.Typ.any ()))
             [exp];
         pexp_attributes =
@@ -171,7 +176,7 @@ let handle_raw_structure loc payload =
           Ast_external_mk.local_external_apply loc ~pval_prim:["#raw_stmt"]
             ~pval_type:
               (Ast_helper.Typ.arrow ~arity:(Some 1)
-                 {lbl = Nolabel; typ = Ast_helper.Typ.any ()}
+                 {attrs = []; lbl = Nolabel; typ = Ast_helper.Typ.any ()}
                  (Ast_helper.Typ.any ()))
             [exp];
       }